# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for reading and writing Bear notes via the hub.",
        "title": "Salmon Hub API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/notes/{noteID}/tags": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds a tag to an existing note and enqueues the action for sync to Bear. Requires Idempotency-Key header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "Add a tag to a note",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note ID",
                        "name": "noteID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Idempotency key for deduplication",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Tag to add",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.addTagRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.WriteQueueItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/tags": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all tags synced from Bear.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "List tags",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tag"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/tags/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Renames a tag across all notes and enqueues the action for sync to Bear. Requires Idempotency-Key header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "Rename a tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Idempotency key for deduplication",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "New tag name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.renameTagRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/models.WriteQueueItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a tag from all notes and enqueues the action for sync to Bear. Requires Idempotency-Key header.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "Delete a tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Idempotency key for deduplication",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/models.WriteQueueItem"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "descriptive error message"
                }
            }
        },
        "api.addTagRequest": {
            "type": "object",
            "properties": {
                "tag": {
                    "type": "string",
                    "example": "work/projects"
                }
            }
        },
...
        "api.renameTagRequest": {
            "type": "object",
            "properties": {
                "new_name": {
                    "type": "string",
                    "example": "work/archived-projects"
                }
            }
        },
...
        "models.Tag": {
            "type": "object",
            "properties": {
                "bear_id": {
                    "type": "string"
                },
                "bear_raw": {
                    "type": "string"
                },
                "encrypted": {
                    "type": "integer"
                },
                "encrypted_at": {
                    "type": "string"
                },
                "hide_subtags_notes": {
                    "type": "integer"
                },
                "hide_subtags_notes_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_root": {
                    "type": "integer"
                },
                "modified_at": {
                    "description": "Dates (ISO 8601)",
                    "type": "string"
                },
                "pinned": {
                    "type": "integer"
                },
                "pinned_at": {
                    "type": "string"
                },
                "pinned_notes_at": {
                    "type": "string"
                },
                "sorting": {
                    "type": "integer"
                },
                "sorting_at": {
                    "type": "string"
                },
                "sorting_direction": {
                    "type": "integer"
                },
                "sorting_direction_at": {
                    "type": "string"
                },
                "tag_con": {
                    "type": "string"
                },
                "tag_con_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "models.WriteQueueItem": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "create | update | add_tag | trash | add_file | archive | rename_tag | delete_tag",
                    "type": "string"
                },
                "applied_at": {
                    "type": "string"
                },
                "consumer_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "lease_until": {
                    "type": "string"
                },
                "note_id": {
                    "type": "string"
                },
                "note_sync_status": {
                    "description": "NoteSyncStatus is populated during lease to indicate the associated note's sync_status.\nNot stored in DB — filled by JOIN during LeaseQueueItems.",
                    "type": "string"
                },
                "payload": {
                    "description": "JSON with operation data",
                    "type": "string"
                },
                "processing_by": {
                    "type": "string"
                },
                "status": {
                    "description": "pending | processing | applied | failed",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Consumer Bearer token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}